rat: Construction of head terms is implemented
Head terms = it is a constructor where arguments are all logic
variables. This is needed to make unification of rational trees finish.

In the implementation we create a bunch of new variables and  extend
substitution with approprite bindings.

TODO: Don't copy terms if the input is already a head term

Signed-off-by: Kakadu <Kakadu@pm.me>